Move TabContentType to a new file
