+++ catowserAndroid/browser-content/src/main/java/com/browser/content/BrowserContentView.kt
@@ -3,15 +3,7 @@ package com.browser.content
 import androidx.compose.material.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.tooling.preview.Preview
-import org.cotton.base.Site
-
-sealed class TabContentType {
-    class Blank(): TabContentType()
-    class TopSites(): TabContentType()
-    class Favorites(): TabContentType()
-    class Homepage(): TabContentType()
-    class SiteContent(val site: Site): TabContentType()
-}
+import com.browser.content.TabContentType
 
 @Composable
 fun BrowserContent(contentType: TabContentType) {
